auto_agent/requirements.txt
# OmniAgent X ULTIMATE - COMPLETE Dependencies

# AI & Brain
openai>=1.3.0

# GUI
customtkinter>=5.2.0

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.1

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Computer Control (Mouse/Keyboard/Screenshot)
pyautogui>=0.9.54
pillow>=10.1.0

# Data Processing
openpyxl>=3.1.0
pandas>=2.0.0

# Environment
python-dotenv>=1.0.0

# ===============================================
# NEW: VOICE SYSTEM
# ===============================================
openai-whisper  # Speech-to-text (eshitish)
edge-tts  # Text-to-speech (gapirish)

# ===============================================
# NEW: TELEGRAM BOT
# ===============================================
python-telegram-bot>=20.0

# ===============================================
# OPTIONAL: MORE POWER
# ===============================================
numpy>=1.24.0
matplotlib>=3.7.0